Move userDeleteWith popup confirm button to own component
